Release 0.2.0
rust-libretro@0.2.0
rust-libretro-example-core@0.2.0
rust-libretro-proc@0.2.0
rust-libretro-sys@0.2.0
rust-libretro-sys-proc@0.2.0

+++ rust-libretro/examples/input/lib.rs
@@ -171,7 +171,7 @@ impl Core for InputTestCore {
 
         let width = fb.width;
         let height = fb.height;
-        let pitch = fb.pitch as u64;
+        let pitch = fb.pitch;
         ctx.draw_frame(data, width, height, pitch);
     }
 }
